feat: event polling for other platforms

+++ engine/include/sf_core/game_types.hpp
@@ -13,8 +13,8 @@ struct GameInstance {
     GameInstance& operator=(const GameInstance& rhs) = delete;
 
     std::function<bool(const GameInstance*)>                        init;
-    std::function<bool(const GameInstance*, f32 delta_time)>        update;
-    std::function<bool(const GameInstance*, f32 delta_time)>        render;
+    std::function<bool(const GameInstance*, f64 delta_time)>        update;
+    std::function<bool(const GameInstance*, f64 delta_time)>        render;
     std::function<void(const GameInstance*, u32 width, u32 height)> on_resize;
     void*                                                           game_state;
     sf::ApplicationConfig                                           app_config;
