+++ testbed/include/game.hpp
@@ -8,6 +8,6 @@ struct GameState {
 };
 
 bool init(const sf::GameInstance* game_inst);
-bool update(const sf::GameInstance* game_inst, f32 delta_time);
-bool render(const sf::GameInstance* game_inst, f32 delta_time);
+bool update(const sf::GameInstance* game_inst, f64 delta_time);
+bool render(const sf::GameInstance* game_inst, f64 delta_time);
 void on_resize(const sf::GameInstance* game_inst, u32 width, u32 height);
